Actual working logic for trimming.
Simpler solution = better solution.

Program initially trims the photo to get rid of random edges on a scan, than we blur it to get rid of small dots on scans and we look for white pixels (on an inverted image). The outter most detected pixels are use to form a smallest bounding box which is used for trimming.